added dynamic metadata generation per page instead of hardcoded values
